requirements.txt
torch==2.1.2
torchvision==0.16.2
timm==1.0.3
scikit-learn==1.5.0
streamlit==1.35.0
pillow==10.2.0
tqdm==4.66.4
numpy==1.26.3